:zap: Added buffer batching

+++ src/engine/Apollo.F1.Compute.Cuda.Native/src/memory.cu
@@ -1,9 +1,7 @@
 #include "memory.cuh"
-#include "stdio.h"
 
-void* allocate_global_memory(int iBytes)
+void* allocate_global_memory(int64_t iBytes)
 {
-   // printf("Allocating %d bytes\n", iBytes);
     void* ptr;
     cudaMalloc(&ptr, iBytes);
     return ptr;
@@ -11,26 +9,25 @@ void* allocate_global_memory(int iBytes)
 
 void destroy_global_memory(void* ptr)
 {
-    //printf("Deallocating bytes\n");
     cudaFree(ptr);
 }
 
-void copy_host_to_device(void* pSrc, void* pDst, int iLength)
+void copy_host_to_device(void* pSrc, void* pDst, int64_t iLength)
 {
     cudaMemcpy(pDst, pSrc, iLength, cudaMemcpyHostToDevice);
 }
 
-void copy_device_to_host(void* pSrc, void* pDst, int iLength)
+void copy_device_to_host(void* pSrc, void* pDst, int64_t iLength)
 {
     cudaMemcpy(pDst, pSrc, iLength, cudaMemcpyDeviceToHost);
 }
 
-void copy_device_to_device(void* pSrc, void* pDst, int iLength)
+void copy_device_to_device(void* pSrc, void* pDst, int64_t iLength)
 {
     cudaMemcpy(pDst, pSrc, iLength, cudaMemcpyDeviceToDevice);
 }
 
-void device_memset(void* pDst, int iLength, int value)
+void device_memset(void* pDst, int64_t iLength, int value)
 {
     cudaMemset(pDst, value, iLength);
 }
