+++ src/engine/Apollo.F1.Compute.Cuda/Buffers/GlobalMemory.cs
@@ -8,22 +8,22 @@ namespace Apollo.F1.Compute.Cuda.Buffers;
 public static class GlobalMemory
 {
     [DllImport(Dll.Name, EntryPoint = "allocate_global_memory", CallingConvention = CallingConvention.Cdecl)]
-    public static extern IntPtr Malloc(int bytes);
+    public static extern IntPtr Malloc(long bytes);
 
     [DllImport(Dll.Name, EntryPoint = "destroy_global_memory", CallingConvention = CallingConvention.Cdecl)]
     public static extern void Free(IntPtr ptr);
 
     [DllImport(Dll.Name, EntryPoint = "copy_host_to_device", CallingConvention = CallingConvention.Cdecl)]
-    public static extern void CopyHostToDevice([MarshalAs(UnmanagedType.LPArray)] double[] src, IntPtr dst, int length);
+    public static extern void CopyHostToDevice([MarshalAs(UnmanagedType.LPArray)] double[] src, IntPtr dst, long length);
     
     [DllImport(Dll.Name, EntryPoint = "copy_device_to_host", CallingConvention = CallingConvention.Cdecl)]
-    public static extern void CopyDeviceToHost(IntPtr src, [MarshalAs(UnmanagedType.LPArray)] double[] dst, int length);
+    public static extern void CopyDeviceToHost(IntPtr src, [MarshalAs(UnmanagedType.LPArray)] double[] dst, long length);
 
     [DllImport(Dll.Name, EntryPoint = "copy_device_to_device", CallingConvention = CallingConvention.Cdecl)]
-    public static extern void CopyDeviceToDevice(IntPtr src, IntPtr dst, int length);
+    public static extern void CopyDeviceToDevice(IntPtr src, IntPtr dst, long length);
 
     [DllImport(Dll.Name, EntryPoint = "device_memset", CallingConvention = CallingConvention.Cdecl)]
-    public static extern void Memset(IntPtr dst, int length, int value);
+    public static extern void Memset(IntPtr dst, long length, int value);
     
     public static IntPtr OffsetOf(IntPtr ptr, int offset)
     {
